refactor(domain): migrate PO/requisition quantity to BigDecimal
PurchaseOrderItem.quantity and RequisitionItem.quantity were
Integer, but other line items in the same domain (InvoiceItem,
GoodReceiptItem, DispatchNoteItem) were already BigDecimal.
That made it impossible to order fractional units (e.g.,
1.5 kg, 0.25 drums) and forced type coercion in every service
method.

- Change both entity fields to BigDecimal(precision=18, scale=4)
  and update unit_price precision the same way.
- Update request/response DTOs: @NotNull + @DecimalMin on the
  request, BigDecimal on the response.
- Drop the BigDecimal.valueOf(poItem.getQuantity()) calls in
  ProcurementService, WarehouseService, and InvoiceService —
  the field is already BigDecimal.
- Flyway V3__quantity_bigdecimal.sql casts the columns
  in place. Existing integer values are preserved as e.g.
  5 -> 5.0000.

Part of Phase 3 (data model) of the audit.

+++ src/main/java/com/visco/backend/models/dtos/PurchaseOrderItemResponse.java
@@ -3,5 +3,5 @@ package com.visco.backend.models.dtos;
 import java.math.BigDecimal;
 
 public record PurchaseOrderItemResponse(Long productId, String productName, String productSku,
-        int quantity, BigDecimal unitPrice, BigDecimal subtotal) {
+        BigDecimal quantity, BigDecimal unitPrice, BigDecimal subtotal) {
 }
