+++ src/main/java/com/visco/backend/models/dtos/RequisitionItemResponse.java
@@ -1,9 +1,11 @@
 package com.visco.backend.models.dtos;
 
+import java.math.BigDecimal;
+
 public record RequisitionItemResponse(
     Long productId,
     String productName,
     String productSku,
-    int quantity,
+    BigDecimal quantity,
     String notes
 ) {}
